pyproject.toml
[tool.poetry]
name = "llmedit"
version = "0.1.0"
description = "The all‑in‑one LLM toolkit for editing, styling, and translating."
authors = ["Oleksandr Kostenko <sanyokkua@gmail.com>"]
readme = "README.md"
include = ["data/prompts/*.txt"]
packages = [
    { include = "llmedit", from = "src" }
]


[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pyqt6 = ">=6.9.1,<7.0.0"
ollama = ">=0.5.1,<0.6.0"
llama-cpp-python = "^0.3.14"
huggingface-hub = "^0.33.4"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pyright = "^1.1.403"
pytest = "^7.0"

[tool.poetry.scripts]
llmedit = "llmedit.main:start_application"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
